+++ packages/builder/src/fee/FeeBuilder.interfaces.ts
@@ -1,25 +1,42 @@
 import type {
   AnyChain,
-  AnyParachain,
+  AssetAmount,
   ChainAsset,
 } from '@moonbeam-network/xcm-types';
 import type { ApiPromise } from '@polkadot/api';
 import type { Enum } from '@polkadot/types';
 import type { StagingXcmV3MultiLocation } from '@polkadot/types/lookup';
 import type { ConfigBuilder } from '../builder.interfaces';
+import type { SubstrateQueryConfig } from '../types';
+import type { ContractConfig } from '../types/evm/ContractConfig';
 import type { SubstrateCallConfig } from '../types/substrate/SubstrateCallConfig';
 
 export type FeeConfigBuilder = ConfigBuilder<
-  SubstrateCallConfig,
+  SubstrateCallConfig | ContractConfig,
   FeeConfigBuilderParams
+>;
+
+export type BridgeFeeConfigBuilder = ConfigBuilder<
+  SubstrateQueryConfig | ContractConfig,
+  BridgeFeeConfigBuilderParams
 >;
 
 export interface FeeConfigBuilderParams {
   address: string;
   api: ApiPromise;
   asset: ChainAsset;
-  destination: AnyParachain;
+  balance?: AssetAmount;
+  destination: AnyChain;
   feeAsset: ChainAsset;
+  source: AnyChain;
+}
+
+export interface BridgeFeeConfigBuilderParams {
+  asset: ChainAsset;
+  feeAsset: ChainAsset;
+  address: string;
+  balance?: AssetAmount;
+  destination: AnyChain;
   source: AnyChain;
 }