+++ packages/builder/src/mrl/providers/snowbridge/extrinsic/index.ts
@@ -1,0 +1,6 @@
+import { ethereumTokenTransfers } from './ethereumTokenTransfer/ethereumTokenTransfers';
+import { xcmPallet } from './xcmPallet/xcmPallet';
+
+export function extrinsic() {
+  return { ethereumTokenTransfers, xcmPallet };
+}