+++ packages/builder/src/types/substrate/SubstrateQueryConfig.ts
@@ -1,7 +1,10 @@
 import { BaseConfig, type BaseConfigConstructorParams } from '../BaseConfig';
 
+type QueryType = 'query' | 'call';
+
 export interface QueryConfigConstructorParams
   extends BaseConfigConstructorParams {
+  queryType?: QueryType;
   args?: unknown[];
   // biome-ignore lint/suspicious/noExplicitAny: not sure how to fix this
   transform: (data: any) => Promise<bigint>;
@@ -9,6 +12,8 @@ export interface QueryConfigConstructorParams
 
 export class SubstrateQueryConfig extends BaseConfig {
   readonly args: unknown[];
+
+  readonly queryType: QueryType;
 
   readonly transform: (data: unknown) => Promise<bigint>;
 
@@ -19,11 +24,13 @@ export class SubstrateQueryConfig extends BaseConfig {
   constructor({
     args = [],
     transform,
+    queryType = 'query',
     ...other
   }: QueryConfigConstructorParams) {
     super({ ...other });
 
     this.args = args;
+    this.queryType = queryType;
     this.transform = transform;
   }
 }